Pull out reducer type constants to local and common folders
