+++ assets/src/edit-story/app/media/common/types.js
@@ -14,4 +14,9 @@
  * limitations under the License.
  */
 
-export const INITIAL_STATE = 'INITIAL_STATE';
+export const FETCH_MEDIA_START = 'FETCH_MEDIA_START';
+export const FETCH_MEDIA_SUCCESS = 'FETCH_MEDIA_SUCCESS';
+export const FETCH_MEDIA_ERROR = 'FETCH_MEDIA_ERROR';
+export const SET_NEXT_PAGE = 'SET_NEXT_PAGE';
+export const UPDATE_MEDIA_ELEMENT = 'UPDATE_MEDIA_ELEMENT';
+export const DELETE_MEDIA_ELEMENT = 'DELETE_MEDIA_ELEMENT';
